Choose real or fake (for tests) SyncAdapter over DI
